db+sqlite+pg: extract shared address converters to addressstore_common
Move the pure address row/secret conversion and validation helpers out of
the monolithic addresses_common.go (db) and addresses.go (sqlite, pg) into
new addressstore_common.go files.
